Add JUnit library and create tests, update Command and CmdInterpreter classes
This commit includes the addition of JUnit4 and JUnit5.8.1 in the module-library, and a creation of a new CmdInterpreterTest class for testing purposes. Several changes are made in Command.java to provide getter methods and implement 'equals' and 'hashCode' methods. Updates in CmdInterpreter.java are made to refine the process of reading command data.